last commit form ds pc

+++ EVM/src/main/java/com/istiaque/EVM/repository/RequestRejectionMesgRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * Created by Mohammad Istiaque Hossain (DS00688) on 30/Oct/2019.
  */
 public interface RequestRejectionMesgRepository extends JpaRepository<RequestRejectionMesg, Integer> {
-    RequestRejectionMesg findByRequestId(Integer requestId);
+    RequestRejectionMesg findByRequestId(Long requestId);
 }
